authorization for simpleList

+++ src/main/java/apps/juice_up/rest/SimpleListResource.java
@@ -1,14 +1,19 @@
 package apps.juice_up.rest;
 
+import apps.juice_up.model.JwtUserDetails;
 import apps.juice_up.model.SimpleListDTO;
 import apps.juice_up.service.SimpleListService;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import jakarta.validation.Valid;
 import java.util.List;
+import java.util.Objects;
+
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 @CrossOrigin(methods = {RequestMethod.GET, RequestMethod.PUT, RequestMethod.DELETE, RequestMethod.POST})
 @RestController
@@ -22,13 +27,18 @@ public class SimpleListResource {
     }
 
     @GetMapping
-    public ResponseEntity<List<SimpleListDTO>> getAllSimpleLists() {
-        return ResponseEntity.ok(simpleListService.findAll());
+    public ResponseEntity<List<SimpleListDTO>> getAllSimpleLists(@AuthenticationPrincipal JwtUserDetails principal) {
+        return ResponseEntity.ok(simpleListService.findByUserId(principal.getId()));
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<SimpleListDTO> getSimpleList(@PathVariable(name = "id") final Long id) {
-        return ResponseEntity.ok(simpleListService.get(id));
+    public ResponseEntity<SimpleListDTO> getSimpleList(
+            @PathVariable(name = "id") final Long id, @AuthenticationPrincipal JwtUserDetails principal) {
+        var simpleList = simpleListService.get(id);
+        if(!Objects.equals(simpleList.getUser(), principal.getId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+        return ResponseEntity.ok(simpleList);
     }
 
     @PostMapping
@@ -40,15 +50,27 @@ public class SimpleListResource {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Void> updateSimpleList(@PathVariable(name = "id") final Long id,
-            @RequestBody @Valid final SimpleListDTO simpleListDTO) {
+    public ResponseEntity<Void> updateSimpleList(
+            @PathVariable(name = "id") final Long id,
+            @RequestBody @Valid final SimpleListDTO simpleListDTO,
+            @AuthenticationPrincipal JwtUserDetails principal
+    ) {
+        var simpleList = simpleListService.get(id);
+        if(!Objects.equals(simpleList.getUser(), principal.getId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         simpleListService.update(id, simpleListDTO);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/{id}")
     @ApiResponse(responseCode = "204")
-    public ResponseEntity<Void> deleteSimpleList(@PathVariable(name = "id") final Long id) {
+    public ResponseEntity<Void> deleteSimpleList(
+            @PathVariable(name = "id") final Long id, @AuthenticationPrincipal JwtUserDetails principal) {
+        var simpleList = simpleListService.get(id);
+        if(!Objects.equals(simpleList.getUser(), principal.getId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         simpleListService.delete(id);
         return ResponseEntity.noContent().build();
     }
